refactor: wrap all jose errors with our own
