+++ .changeset/quick-socks-whisper.md
@@ -1,0 +1,5 @@
+---
+"@labdigital/federated-token": minor
+---
+
+Delete refreshToken if invalid
